requirements.txt
aiogram>=3.7.0
openpyxl
python-dotenv
